+++ models/getContactById.js
@@ -1,0 +1,12 @@
+const listContacts = require("./listContacts");
+
+const getContactById = async (contactId) => {
+  const contacts = await listContacts();
+  const result = contacts.find((item) => item.id === contactId);
+  if (!result) {
+    return null;
+  }
+  return result;
+};
+
+module.exports = getContactById;
